+++ Abs/Entities/Entity.cpp
@@ -1,30 +1,30 @@
 #include <iostream>
 #include "Entity.h"
 
-Abs::Entity::Entity(float x, float y, float w, float h) {
+Abs::Entity::Entity(double x, double y, double w, double h) {
     xPos = x;
     yPos = y;
     width = w;
     height = h;
 }
 
-float Abs::Entity::getXPos() {
+double Abs::Entity::getXPos() {
     return xPos;
 }
 
-float Abs::Entity::getYPos() {
+double Abs::Entity::getYPos() {
     return yPos;
 }
 
-float Abs::Entity::getWidth() const {
+double Abs::Entity::getWidth() const {
     return width;
 }
 
-float Abs::Entity::getHeight() const {
+double Abs::Entity::getHeight() const {
     return height;
 }
 
-void Abs::Entity::move(float dx, float dy) {
+void Abs::Entity::move(double dx, double dy) {
     if ((xPos + dx) < (width/2))
         xPos = (width/2);
     else if (xPos + dx + (width/2) > 1)
