+++ SDL/Entities/SDLPlayerBullet.cpp
@@ -1,6 +1,6 @@
 #include "SDLPlayerBullet.h"
 
-Sdl::SDLPlayerBullet::SDLPlayerBullet(Sdl::SDLWindow* window, float x, float y, float w, float h) : Abs::PlayerBullet(x, y, w, h) {
+Sdl::SDLPlayerBullet::SDLPlayerBullet(Sdl::SDLWindow* window, double x, double y, double w, double h) : Abs::PlayerBullet(x, y, w, h) {
     this->window = window;
 }
