+++ SDL/SDLFactory.cpp
@@ -11,11 +11,11 @@ Sdl::SDLFactory::~SDLFactory() {
     delete window;
 }
 
-Abs::PlayerShip* Sdl::SDLFactory::createPlayerShip(float x, float y, float w, float h) {
+Abs::PlayerShip* Sdl::SDLFactory::createPlayerShip(double x, double y, double w, double h) {
     return new Sdl::SDLPlayerShip(window, x, y, w, h);
 }
 
-Abs::EnemyShip* Sdl::SDLFactory::createEnemyShip(float x, float y, float w, float h, ENTITY e) {
+Abs::EnemyShip* Sdl::SDLFactory::createEnemyShip(double x, double y, double w, double h, ENTITY e) {
     return new Sdl::SDLEnemyShip(window, x, y, w, h, e);
 }
 
@@ -23,21 +23,21 @@ Abs::Controller* Sdl::SDLFactory::createController() {
     return new Sdl::SDLController();
 }
 
-Abs::PlayerBullet* Sdl::SDLFactory::createPlayerBullet(float x, float y, float w, float h) {
+Abs::PlayerBullet* Sdl::SDLFactory::createPlayerBullet(double x, double y, double w, double h) {
     window->playSound(SHOOT);
     return new Sdl::SDLPlayerBullet(window, x, y, w, h);
 }
 
-Abs::EnemyBullet* Sdl::SDLFactory::createEnemyBullet(float x, float y, float w, float h) {
+Abs::EnemyBullet* Sdl::SDLFactory::createEnemyBullet(double x, double y, double w, double h) {
     return new Sdl::SDLEnemyBullet(window, x, y, w, h);
 }
 
-Abs::PosBonus* Sdl::SDLFactory::createPosBonus(float x, float y, float w, float h) {
+Abs::PosBonus* Sdl::SDLFactory::createPosBonus(double x, double y, double w, double h) {
     window->playSound(BONUS_SPAWN);
     return new Sdl::SDLPosBonus(window, x, y, w, h);
 }
 
-Abs::NegBonus* Sdl::SDLFactory::createNegBonus(float x, float y, float w, float h) {
+Abs::NegBonus* Sdl::SDLFactory::createNegBonus(double x, double y, double w, double h) {
     window->playSound(BONUS_SPAWN);
     return new Sdl::SDLNegBonus(window, x, y, w, h);
 }
